+++ packages/database/CHANGELOG.md
@@ -1,5 +1,101 @@
 # @nospipi/database
 
+## 6.0.1
+
+### Patch Changes
+
+- updates
+
+## 6.0.0
+
+### Major Changes
+
+- updates
+- add sales channels schema
+- updates
+
+### Patch Changes
+
+- updatess
+- none
+- none
+- none
+- none
+- none
+- dgfhdgfh
+- none
+
+## 5.0.0
+
+### Major Changes
+
+- updates
+- add sales channels schema
+- updates
+
+### Patch Changes
+
+- none
+- none
+- none
+- none
+- none
+- dgfhdgfh
+- none
+
+## 4.0.0
+
+### Major Changes
+
+- updates
+- add sales channels schema
+- updates
+
+### Patch Changes
+
+- none
+- none
+- none
+- none
+- none
+- dgfhdgfh
+- none
+
+## 3.0.0
+
+### Major Changes
+
+- updates
+- add sales channels schema
+- updates
+
+### Patch Changes
+
+- none
+- none
+- none
+- none
+- none
+- dgfhdgfh
+- none
+
+## 2.0.0
+
+### Major Changes
+
+- add sales channels schema
+- updates
+
+### Patch Changes
+
+- none
+- none
+- none
+- none
+- none
+- dgfhdgfh
+- none
+
 ## 1.0.0
 
 ### Major Changes